requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
idna==3.6
igraph==0.11.3
pipreqs==0.4.13
python-igraph==0.11.3
PyYAML==6.0.1
requests==2.31.0
tabulate==0.9.0
texttable==1.7.0
urllib3==2.1.0
yarg==0.1.9
